Add currentscavengerhunt app. Remove currentscavengerhunt test from scavengerhunts
