+++ scavengerhunts/tests.py
@@ -230,27 +230,3 @@ class ScavengerHuntEndpointTestCase(TestCase):
         self.assertEqual(response.json()[1]['description'], "second_description")
         self.assertEqual(response.json()[1]['user_id'], self.user2.id)
         self.assertEqual(response.json()[1]['username'], self.user2.username)
-
-
-class CurrentScavengerHuntEndpointTestCase(TestCase):
-
-    def setUp(self):
-        self.user = User.objects.create(username='testuser', email='test@email.com', password='password')
-        self.token = Token.objects.get(user=self.user)
-
-        self.client = APIClient()
-        self.client.credentials(HTTP_AUTHORIZATION='Token ' + self.token.key)
-
-        self.scavenger_hunt = ScavengerHunt.objects.create(name="scavenger hunt", description="tour denver's best speakeasies")
-
-
-    def test_scavenger_hunt_create_endpoint(self):
-        data = { 'scavenger_hunt_id': 1 }
-
-        response = self.client.post('/api/v1/current_scavenger_hunts/', data, format='json')
-
-        self.assertEqual(response.status_code, 201)
-
-        response = self.client.post('/api/v1/current_scavenger_hunts/', data, format='json')
-
-        self.assertEqual(response.status_code, 200)
